reordered install scripts and added install flags.

+++ install/scripts/120_ardupilot.sh
@@ -7,6 +7,14 @@ SCRIPTS_DIR="/home/pi/drone-rpi3-bookworm-32bit/install/scripts"
 source "$SCRIPTS_DIR/00_common.env"
 source "$SCRIPTS_DIR/00_lib.sh"
 
+INSTALL_FLAG="$LOG_DIR/ardupilot"
+
+log "checking to see if previous install ran successfully..."
+if [ -f $INSTALL_FLAG]; then 
+    log "ArduPilot install was already run successfully..."
+    return 0 
+fi 
+
 log "getting the code..."
 if [ -d "$ARDUPILOT_DIR" ]; then 
     rm -rf "$ARDUPILOT_DIR"
@@ -19,7 +27,7 @@ cd $ARDUPILOT_DIR
 git submodule update --init --recursive
 
 log "installing build tools..."
-sudo apt install -y python3-serial python3-dev libtool libxml2-dev libxslt1-dev \
+sudo apt install -y -qq python3-serial python3-dev libtool libxml2-dev libxslt1-dev \
   gawk python3-pip pkg-config build-essential ccache libffi-dev \
   libjpeg-dev zlib1g-dev python3-empy python3-pexpect
 sudo python3 -m pip install --upgrade pymavlink MAVProxy --break-system-packages
@@ -30,3 +38,5 @@ log "building ArduCopter for Navio2..."
 
 log "adding ArduPilot as a service..."
 sudo bash "$SCRIPTS_DIR/131_ardupilot_service.sh"
+
+touch "$INSTALL_FLAG"
